website/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.43.0
Django==5.1.7
django-cloudinary-storage==0.3.0
idna==3.10
ipywidgets==8.1.2
jupyterlab_widgets==3.0.10
matplotlib==3.8.4
numpy==1.26.4
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.66.4
urllib3==2.3.0
widgetsnbextension==4.0.10

gunicorn
whitenoise
psycopg2-binary